requirements.txt
pandas==2.2.1
numpy==1.26.4
statsmodels==0.14.1
scikit-learn==1.4.1.post1
google-generativeai==0.4.1
python-dotenv==1.0.1
pytest==8.1.1
streamlit==1.32.0
altair==5.2.0
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.3
sqlalchemy==2.0.27
